rename capture_backward_subgraph_guard
