+++ python/paddle/utils/__init__.py
@@ -22,6 +22,7 @@ from . import (  # noqa: F401
     layers_utils,
     unique_name,
 )
+from .bwd_graph_utils import capture_backward_subgraph_guard  # noqa: F401
 from .deprecated import deprecated
 from .environments import strtobool as strtobool
 from .fwd_graph_utils import capture_forward_subgraph_guard  # noqa: F401
