+++ test/custom_op/test_custom_relu_op_setup.py
@@ -344,9 +344,7 @@ class TestNewCustomOpSetUpInstall(unittest.TestCase):
         )
         with (
             paddle.utils.capture_forward_subgraph_guard("./tmp_subgraph"),
-            paddle.base.framework.capture_backward_subgraph_guard(
-                "./tmp_debug_info"
-            ),
+            paddle.utils.capture_backward_subgraph_guard("./tmp_debug_info"),
         ):
             x = paddle.randn([5, 5])
             x.stop_gradient = False
